ProcedureDeclarationCheckerTest and PublicMethodsHaveTypeCheckerTest are mixed.
